Fixed database index mapping issue. Added List manager list name to its Item list screen Added back icon to Item list screen and on click navigates back to previous screen

+++ app/src/main/java/com/shoppinglist/app/Item.java
@@ -8,13 +8,13 @@ import android.os.Parcelable;
  */
 public class Item implements Parcelable {
     private int id;
-    private int array_id;
+    private long array_id;
     private String name;
     private String description;
 
     public Item(){}
 
-    public Item(int id, int array_id, String name, String description) {
+    public Item(int id, long array_id, String name, String description) {
         this.id = id;
         this.array_id = array_id;
         this.name = name;
@@ -26,7 +26,7 @@ public class Item implements Parcelable {
         this.description = description;
     }
 
-    public Item(int array_id, String name, String description) {
+    public Item(long array_id, String name, String description) {
         this.array_id = array_id;
         this.name = name;
         this.description = description;
@@ -40,11 +40,11 @@ public class Item implements Parcelable {
 
     public int getId() { return id; }
 
-    public int getArrayId() { return array_id; }
+    public long getArrayId() { return array_id; }
 
     public void setId(int id) { this.id = id; }
 
-    public void setArrayId(int array_id) { this.array_id = array_id; }
+    public void setArrayId(long array_id) { this.array_id = array_id; }
 
     public String getName() {
         return name;
